Catalogue controller. Remove base fragment classes
