+++ app/src/main/java/eu/kanade/tachiyomi/ui/base/activity/BaseRxActivity.kt
@@ -1,7 +1,5 @@
 package eu.kanade.tachiyomi.ui.base.activity
 
-import android.os.Bundle
-import eu.kanade.tachiyomi.App
 import eu.kanade.tachiyomi.ui.base.presenter.BasePresenter
 import eu.kanade.tachiyomi.util.LocaleHelper
 import nucleus.view.NucleusAppCompatActivity
@@ -12,17 +10,6 @@ abstract class BaseRxActivity<P : BasePresenter<*>> : NucleusAppCompatActivity<P
 
     init {
         LocaleHelper.updateConfiguration(this)
-    }
-
-    override fun onCreate(savedState: Bundle?) {
-        val superFactory = presenterFactory
-        setPresenterFactory {
-            superFactory.createPresenter().apply {
-                val app = application as App
-                context = app.applicationContext
-            }
-        }
-        super.onCreate(savedState)
     }
 
     override fun getActivity() = this
