+++ app/src/main/java/eu/kanade/tachiyomi/ui/base/controller/SecondaryDrawerController.kt
@@ -1,0 +1,11 @@
+package eu.kanade.tachiyomi.ui.base.controller
+
+import android.support.v4.widget.DrawerLayout
+import android.view.ViewGroup
+
+interface SecondaryDrawerController {
+
+    fun createSecondaryDrawer(drawer: DrawerLayout): ViewGroup
+
+    fun cleanupSecondaryDrawer(drawer: DrawerLayout)
+}
