--- app/src/main/java/eu/kanade/tachiyomi/ui/base/fragment/BaseFragment.kt
@@ -1,7 +1,0 @@
-package eu.kanade.tachiyomi.ui.base.fragment
-
-import android.support.v4.app.Fragment
-
-abstract class BaseFragment : Fragment(), FragmentMixin {
-
-}
