+++ app/src/main/java/eu/kanade/tachiyomi/ui/latest_updates/LatestUpdatesController.kt
@@ -1,0 +1,35 @@
+package eu.kanade.tachiyomi.ui.latest_updates
+
+import android.support.v4.widget.DrawerLayout
+import android.view.Menu
+import android.view.ViewGroup
+import android.widget.FrameLayout
+import eu.kanade.tachiyomi.R
+import eu.kanade.tachiyomi.ui.catalogue.CatalogueController
+import eu.kanade.tachiyomi.ui.catalogue.CataloguePresenter
+
+/**
+ * Fragment that shows the manga from the catalogue. Inherit CatalogueFragment.
+ */
+class LatestUpdatesController : CatalogueController() {
+
+    override fun createPresenter(): CataloguePresenter {
+        return LatestUpdatesPresenter()
+    }
+
+    override fun onPrepareOptionsMenu(menu: Menu) {
+        super.onPrepareOptionsMenu(menu)
+        menu.findItem(R.id.action_search).isVisible = false
+        menu.findItem(R.id.action_set_filter).isVisible = false
+
+    }
+
+    override fun createSecondaryDrawer(drawer: DrawerLayout): ViewGroup {
+        return FrameLayout(drawer.context)
+    }
+
+    override fun cleanupSecondaryDrawer(drawer: DrawerLayout) {
+
+    }
+
+}
