+++ app/src/main/res/layout/fragment_manga_info.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="eu.kanade.tachiyomi.ui.catalogue.CatalogueFragment"
+    tools:context="eu.kanade.tachiyomi.ui.catalogue.CatalogueController"
     android:id="@id/swipe_refresh"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
